Refactor modules, variables, classes, functions
Replaced hard to understand  abbreviations
Updated module names
Updated functions names
Moved methods from current_user to raw_config and raw_registry

In hope of improving the readability of the code

Modules renamed:
- data/data_fields.py -> data/metadata_classes.py\
- gui/tab_subsearch.py -> gui/tab_application_settings.py\
- gui/tab_language.py -> gui/tab_languages.py\
- gui/tab_download.py -> gui/tab_manual_download.py\
- gui/tab_search.py -> gui/tab_search_settings.py

Modules deleted:
- utils/current_user.py
