+++ src/subsearch/utils/raw_config.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from typing import Any, Union
 
 from subsearch.data import __data__
-from subsearch.data.data_fields import UserData
+from subsearch.data.metadata_classes import ApplicationSettings
 
 
 def set_config_key_value(key: str, value: Union[str, int, bool]) -> None:
@@ -45,7 +45,7 @@ def get_config_key(key: str) -> Any:
 
         - GUI stuff
 
-        "cm_icon: bool", "show_download_window: bool",
+        "cm_icon: bool", "manual_download_tab: bool",
         "show_terminal: bool", "file_ext: dict"
 
     Returns:
@@ -55,14 +55,14 @@ def get_config_key(key: str) -> Any:
         "current_language": get_config()["current_language"],
         "languages": get_config()["languages"],
         "subtitle_type": get_config()["subtitle_type"],
-        "percentage": get_config()["percentage"],
+        "percentage_threshold": get_config()["percentage_threshold"],
         "rename_best_match": get_config()["rename_best_match"],
         "context_menu_icon": get_config()["context_menu_icon"],
-        "show_download_window": get_config()["show_download_window"],
-        "threading": get_config()["threading"],
+        "manual_download_tab": get_config()["manual_download_tab"],
+        "use_threading": get_config()["use_threading"],
         "show_terminal": get_config()["show_terminal"],
         "log_to_file": get_config()["log_to_file"],
-        "file_ext": get_config()["file_ext"],
+        "file_extensions": get_config()["file_extensions"],
         "providers": get_config()["providers"],
     }
     return config_json_dict[f"{key}"]
@@ -73,14 +73,14 @@ def set_default_json() -> None:
     data = get_config()
     data["current_language"] = "English"
     data["subtitle_type"] = dict.fromkeys(data["subtitle_type"], True)
-    data["percentage"] = 90
+    data["percentage_threshold"] = 90
     data["rename_best_match"] = True
     data["context_menu_icon"] = True
-    data["show_download_window"] = True
-    data["threading"] = False
+    data["manual_download_tab"] = True
+    data["use_threading"] = False
     data["show_terminal"] = False
     data["log_to_file"] = False
-    data["file_ext"] = dict.fromkeys(data["file_ext"], True)
+    data["file_extensions"] = dict.fromkeys(data["file_extensions"], True)
     data["providers"] = dict.fromkeys(data["providers"], True)
     config_file = Path(__data__) / "config.json"
     with open(config_file, "r+", encoding="utf-8") as file:
@@ -89,13 +89,13 @@ def set_default_json() -> None:
         file.truncate()
 
 
-def get_user_data() -> UserData:
+def get_user_data() -> ApplicationSettings:
     config_file = Path(__data__) / "config.json"
     with open(config_file, encoding="utf-8") as file:
         data = json.load(file)
-    user_data = UserData(
+    user_data = ApplicationSettings(
         **data,
-        language_code3=data["languages"][data["current_language"]],
+        language_iso_639_3=data["languages"][data["current_language"]],
         hearing_impaired=data["subtitle_type"]["hearing_impaired"],
         non_hearing_impaired=data["subtitle_type"]["non_hearing_impaired"],
     )
